bot/requirements.txt
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyserial==3.4
pytest==6.2.3
python-dateutil==2.8.1
python-dotenv==0.17.0
pytz==2021.1
pytzdata==2020.1
requests-unixsocket==0.2.0
pyttsx3==2.90
textblob==0.15.3
twitchrealtimehandler==0.0.2
pydub==0.25.1
fake-useragent==0.1.11